+++ yandex/logging.go
@@ -1,0 +1,10 @@
+package yandex
+
+import (
+	"log"
+)
+
+func init() {
+	// Suppress anything but the log message itself.
+	log.SetFlags(0)
+}